Fix issue with Istio init timing
Fixes: 94

A user reported an error about a missing Istio CRD and it
appears that one must wait for at least three bespoke init
jobs to complete before attempting to do the installation.
With KinD, on my home Internet connection this took several
minutes to run.

Technique sourced from Flagger / @stefanprodan.

In the future, when arkade uses istioctl, hopefully all this
bespoke code won't be required for an unattended installation.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>