Add a GetNamespace and CreateNamespace command
This commit sets a global flag of namespace onto the install command, it
then creates GetNamespace and CreateNamespace functions to allow a
single place to get and create namespaces.

The new style chart_app runs create namespace for you now, on an
existing namespace the function prints a namespace exists message.

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>