Add `arkade install openfaas` integration test
This integration test allows to be sure that `arkade install openfaas`
works as intended: by default, creates 7 deployments inside the cluster
defined in `KUBECONFIG`.

For this test, we spin up a k3s cluster inside a docker container. This
part is handled by Gnomock (https://github.com/orlangure/gnomock). When
the cluster is up and running, we save its kubeconfig into a temporary
file and execute `arkade install openfaas` against it.  To verify that
everything worked, we list existing deployments in `openfaas` namespace
and make sure their names match our expectations.

This test can be further extended by using all supported flags. This
initial implementation is only a baseline, but it is a valid test. Other
apps can be tested as well.

Signed-off-by: Yury Fedorov <orlangure@gmail.com>