Add a WithSecrets to App installer
This converts OpenFaaS to the new app style and adds a .WithSecret() to
app options. This can be used to create a basic secret for now (the
default k8s type of generic) - It could be extended in the future to
create other k8s secret types.

Tested by installing OpenFaaS On a k3d cluster and logging in to the UI
(thich uses the basic auth secret)

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>