Switch to jFrog's Chart Center mirror for stable repos
The stable repos have been removed from their original
GCS bucket. The mirror is giving a 404 error despite the
index.yaml file showing the download URL.

Chart center from jFrog appears to have valid mirrors of the
older charts: https://chartcenter.io

This is not a fix for the issue, but is a workaround.
Apps will have to be removed if no suitable upstream chart is
available going forward.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>