Add support for openfaas pro license to OpenFaaS
This adds support for the new OpenFaaS Pro license to the openfaas app.
The license must be passed as a file to avoid it being printed to the
console.

When set the idler will be installed and have read-only mode set to true
(not making changes). To enable the idler to scale down to 0 you set
`--set faasIdler.readOnly=false`

This has been tested with idler on, off and enabled

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>