fix: support ingress on k8s 1.22+
Add utility to retrieve the list of supported apis from the current k8s
context.

Use this to pass a flag to the ingress creation for the openfaas ingress
app. This allows the app to install with the newer networking.k8s.io/v1
api instead of the deprecated extensions/v1 api that will be removed in
k8s 1.22+ (July 2021)

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>