Automatic generation of README catalogs
The `README.md` tables of the catalog of CLIs and apps might
become out of sync from time to time.
It should be automatically generated and fail build in case not committed.

Using `cog` https://nedbatchelder.com/code/cog/ , the `README.md` file
can be generated and overwrite itself by calling

- `arkade install --print-table`
- `arkade get --output markdown`

The Github Action workflows were updated accordingly to run the
`make lint` as part of the job, in order to verify calling `make docs`
does not make any changes to the repoisotry that weren't commited to
git.

Add `${PWD}/${BIN}` to the Makefile in order to use the latest `arkade`
that was generated by `make dist`.

Bugfix `arkade install --print-table` to be key ordered (otherwise
everytime it produces a different table).

Signed-off-by: Yuval Goldberg <yuvigoldi@gmail.com>