Download binaries to separate temp directory
Fixes #911 where files were downloaded to the same directory
with permissions that prevented another user from running
arkade get.

The changes remove the concept of stash which I don't think
was used or useful and replaces it with a default path for
the current behaviour or --path to match the way that system
installers work today.

Tested with a single download and multiple downloads via
arkade get with and without a custom path. The files were
executable by the owner as expected.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>