Initial version of an update-sdk
This package is designed to work with faas-cli, arkade, k3sup,
inletsctl, etc. which are maintained by a common group.

It may require changes, or some replacement implementations of
the interfaces to work for other CLI tools.

Tested by setting a version via ldflags, to simulate a version
match and a different version, in both cases I saw the correct
result.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>