helm: add `arkade chart bump` to bump chart versions
Add a new command `arkade chart bump --dir ./chart` to bump the minor
of a Helm chart version. To bump multiple charts at once, specify the
parent directory with the `--recursive` flag. By default, the changes
are written to stdout. Specifying the `--write` flag writes the
changes to the respective file.

Signed-off-by: Sanskar Jaiswal <jaiswalsanskar078@gmail.com>

njndjnd