add k8s version strategy
Although there is a version strategy for github, whereby the latest version is resolved via the gh release pages something similar did not exist for k8s releases.

https://cdn.dl.k8s.io/release/stable.txt is made available to enable resolution of the latest version.

This change integrates functionality that interrogates this file and uses its contents to assert the version of `kubectl` to retrieve.  This will negate the situation described in #1048 where the version is out of date until the `arkade` code is updated.

Signed-off-by: Richard Gee <richard@technologee.co.uk>