Fix issue which saw versions resolved twice
The version of a tool was being looked up twice - once prior to downloading
and again post downloading.  This was because GetUrl was being reached twice,
initially to determine the download url and then again to determine whether
the download was an archive.

This change introduces a new function which takes the download url and
uses that to determine whether its an archive.  This removes the additional
round trip and the associated output.  The new function is also covered by
an additional test.

Signed-off-by: Richard Gee <richard@technologee.co.uk>