fix chart upgrade to handle v prefix
It was noted that after a chart upgrade if the upgrade command was
run for a second time it would update two images within openfaas/values.yaml.

This was due to the prom images which are prefixed by v.  This case  was
being handled after the check of whether the latestTag != tag.

This change moves the handling of this case prior to the if statement.

Signed-off-by: Richard Gee <richard@technologee.co.uk>