Move upgradeable check to using semVer/pre
The number of tag version edge cases requiring exception from being upgraded
was increasing to the point where constantly adding to the exception list was
no longer a viable approach.

This change alters how the upgradability of a tag is assessed by parsing the
current and latest tags into SemVers and then assessing the `pre` field of the
version to determine whether either of the tags are a special case.  In order
to achieve this it is necessary to move `latest` out to be a case all of itself
as this would not be nicely handled by NewVersion().

Signed-off-by: Richard Gee <richard@technologee.co.uk>