refactor: consolidate version finding across version strategies
Initially there was a single version strategy of 'github' which meant the
location was standardised.  Then came k8s as a version strategy which took
a different approach and so needed a different implementation.  Later came
Gitlab which required another implementation.

This change consolidates the various release finding approaches into a single
function, which varied based on attributes held about the release 'platform'.

As a demonstration of the benefit of the approach GitLab has been added as a
release platform and the GitLab CLI which is hosted on gitlab has been added
as a tool - superseding #1122

Signed-off-by: Richard Gee <richard@technologee.co.uk>