fix: mitigate effect on tags of semVer
a chart upgrade which contained postgres:16 highlighted that the current appraoch to parsing tags meant that 16 moved to 17.0.0 which doesnt exist. We can alleviate this by using .Original() from semVer once the most recent tag has been found.  Using this also removes the need to manipulate the resulting string in the event of the image using a leading v

Signed-off-by: Richard Gee <richard@technologee.co.uk>