feat: add `ark update readme` command
Adds a subcommand to update that will update the README tables in situ.
This will be useful to contributors who are adding system installs / apps / tools
as it will remove the need for the existing manual mechanism to update the
appropriate tables in the readme.  Running `arkade update readme` will update
each of the three tables in the readme, ready to be committed with their change.

Signed-off-by: Richard Gee <richard@technologee.co.uk>