Add way to hold/ignore images when upgrading compose/charts
When upgrading image versions in a compose file or helm chart
there are cases where certain components like a database version
should be static and not change.

This change introduces a more generic concept than previously
planned of an arkade config file. The ignore section includes
paths to images which should be ignored and filtered out
prior to looking up new images.

Tested with the OpenFaaS pro-builder chart, with the path
buildkit.image, I was able to ignore only that image whilst
having the other images inspected/updated.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>